Added ebook2cw version readout, simplified executable folder management, small optimizations.

+++ ebook2cw.py
@@ -1,13 +1,14 @@
 import helpers
 import os
 import platform
+import subprocess
 
 
 class Ebook2Cw:
-    def __init__(self, base_path, data_folder):
+    def __init__(self, executable_folder):
         """Class initialization.
             Setup internal data and detect if running on supported OS.
-            Currently supported platforms: Windows, Linux.
+            Supported OSes: Windows, Linux.
         """
 
         BASE_URL = 'https://fkurz.net/ham/ebook2cw/'
@@ -17,13 +18,18 @@ class Ebook2Cw:
         self.is_os_supported = False
 
         # assembly paths and urls
+        ebook2cw_folder = os.path.normpath(executable_folder)
+        if os.path.isfile(executable_folder):
+            ebook2cw_folder = os.path.normpath(
+                os.path.dirname(os.path.abspath(executable_folder)))
+
         self.executable_url = BASE_URL + EXECUTABLE_BASE_NAME
         self.executable_local_path = os.path.normpath(os.path.join(
-            base_path, data_folder, EXECUTABLE_BASE_NAME))
+            ebook2cw_folder, EXECUTABLE_BASE_NAME))
 
         self.hash_file_url = BASE_URL + HASH_FILE_NAME
         self.hash_file_local_path = os.path.normpath(os.path.join(
-            base_path, data_folder, HASH_FILE_NAME))
+            ebook2cw_folder, HASH_FILE_NAME))
 
         current_os = platform.system()
 
@@ -55,34 +61,20 @@ class Ebook2Cw:
             if os.path.exists(self.executable_local_path) and os.path.exists(self.hash_file_local_path):
                 executable_file_name = os.path.basename(
                     self.executable_local_path)
+                executable_md5 = helpers.md5(self.executable_local_path)
 
                 with open(os.path.normpath(self.hash_file_local_path), mode="r") as md5_file:
                     for line in md5_file:
                         md5_data = line.strip().split(None, 1)
 
                         # verify calculated and provided hash
-                        executable_md5 = helpers.md5(
-                            self.executable_local_path)
                         if executable_file_name == md5_data[1] and executable_md5 == md5_data[0]:
                             is_verified = True
                             break
 
         return is_verified
 
-    def os_supported(self):
-        """Returns if current OS is supported.
-            If not all fuctionalities will fail.
-
-        Args:
-            None
-
-        Returns:
-            bool: True when OS is supported, False otherwise
-        """
-
-        return self.is_os_supported
-
-    def is_executable_present(self, file_path):
+    def _is_executable_present(self):
         """Returns if ebook2cw executable is available in default directory.
 
         Args:
@@ -101,35 +93,66 @@ class Ebook2Cw:
 
         return is_executable_available
 
-    def get_executable(self):
-        """Downloads latest version of the ebook2cw (OS speciffic)
-            performing download integrity check by MD5 verification.
-            In case of not supported OS download is not performed.
+    def os_supported(self):
+        """Returns if current OS is supported.
+            If not all fuctionalities will fail.
 
         Args:
             None
 
         Returns:
-            bool: True when executable was downloaded and verified,
-                   False otherwise
+            bool: True when OS is supported, False otherwise
         """
 
-        if_downloaded_ok = False
+        return self.is_os_supported
+
+    def get_executable_version(self):
+        version = '0'
+
+        if self.is_os_supported and self._is_executable_present():
+            output = subprocess.Popen(
+                [self.executable_local_path, "-help"], stdout=subprocess.PIPE).communicate()[0].decode()
+            # correct response is: file_name version
+            extracted_version = output.split('\n', 1)[0].split()[1]
+            # basic check for version length which is at least X.Y.Z
+            if len(extracted_version) >= 5:
+                version = extracted_version
+
+        return version
+
+    def get_executable(self, force_latest=True):
+        """Downloads latest version of the ebook2cw (OS speciffic)
+            performing download integrity check by MD5 verification.
+            When not forced it checks if executable exist locally.
+            In case of not supported OS none of above is not performed.
+
+        Args:
+            force_latest (bool): Forces to download latest version
+
+        Returns:
+            bool: True when executable is available, False otherwise
+        """
+
+        if_got_executable = False
 
         if self.is_os_supported:
-            # download files (executable + md5)
-            helpers.get_file_from_web(
-                self.executable_url, self.executable_local_path)
-            helpers.get_file_from_web(
-                self.hash_file_url, self.hash_file_local_path)
+            # when not forced verify if downloaded already
+            if not force_latest and os.path.exists(self.executable_local_path):
+                if_got_executable = True
+            else:
+                # download files (executable + md5)
+                helpers.get_file_from_web(
+                    self.executable_url, self.executable_local_path)
+                helpers.get_file_from_web(
+                    self.hash_file_url, self.hash_file_local_path)
 
-            # verify if both files were downloaded
-            if os.path.exists(self.executable_local_path) and os.path.exists(self.hash_file_local_path):
-                # verify executable's integrity
-                if self._verify_executable_against_md5_file():
-                    if_downloaded_ok = True
-                else:
-                    os.remove(self.executable_local_path)
-                    os.remove(self.hash_file_local_path)
+                if os.path.exists(self.executable_local_path) and os.path.exists(self.hash_file_local_path):
+                    # verify executable's integrity
+                    if self._verify_executable_against_md5_file():
+                        if_got_executable = True
+                    else:
+                        # md5 mismatch - remove files
+                        os.remove(self.executable_local_path)
+                        os.remove(self.hash_file_local_path)
 
-        return if_downloaded_ok
+        return if_got_executable
